feat: add Shiki syntax highlighting to chat markdown
- Extract Shiki singleton + curated language allowlist into shiki-core.ts
  so CodeViewer (skill file viewer) and the new Streamdown plugin share
  one highlighter instance.
- Add custom CodeHighlighterPlugin (shiki-streamdown-plugin.ts) that
  bridges to shiki-core. Avoids the ~9.5MB bundle bloat that the
  official @streamdown/code package would have introduced.
- Override Streamdown's default code-block layout via markdown.css to
  match the docs example: tighter padding, single-row header with
  hover-only action buttons, light/dark Shiki colors via CSS vars.
- Bundle size unchanged: 14.46 MB.
