+++ src/components/chat/markdown.css
@@ -1,0 +1,107 @@
+/* 
+ * Streamdown code block overrides to match the official Streamdown docs layout.
+ * Removes the aggressive prose overrides and tightens up the grid/flex alignment
+ * so the header and actions sit on the same row.
+ */
+
+[data-streamdown="code-block"] {
+  /* Subtle gray wrapper like the example */
+  background: var(--muted);
+  border: 1px solid var(--border);
+  padding: 0.5rem; /* Reduced for a tighter frame */
+  gap: 0;
+  border-radius: 0.5rem;
+  overflow: visible;
+  position: relative;
+}
+
+/* We restore Streamdown's default header but tighten it */
+[data-streamdown="code-block-header"] {
+  display: flex !important;
+  align-items: center;
+  height: 2rem !important;
+  padding: 0 0.5rem;
+  margin-bottom: 0.25rem;
+}
+
+[data-streamdown="code-block-header"] > span {
+  font-family: var(--font-mono);
+  font-size: 0.75rem;
+  color: var(--muted-foreground);
+  text-transform: lowercase;
+}
+
+/* Position the action row inside the header area */
+[data-streamdown="code-block"] > div.sticky {
+  position: absolute !important;
+  top: 0.5rem !important; 
+  right: 0.5rem !important;
+  margin-top: 0 !important;
+  height: 2rem !important;
+  display: flex !important;
+  align-items: center;
+  justify-content: flex-end;
+  pointer-events: none !important;
+  z-index: 20;
+}
+
+[data-streamdown="code-block-actions"] {
+  /* Streamdown puts a background on the action group, let's remove it for the clean look */
+  background: transparent !important;
+  border: none !important;
+  gap: 0.25rem;
+  opacity: 0;
+  transition: opacity 0.2s;
+  pointer-events: auto !important;
+}
+
+[data-streamdown="code-block"]:hover [data-streamdown="code-block-actions"] {
+  opacity: 1;
+}
+
+[data-streamdown="code-block-actions"] button {
+  padding: 0.25rem !important;
+  color: var(--muted-foreground);
+}
+
+[data-streamdown="code-block-actions"] button:hover {
+  color: var(--foreground);
+}
+
+[data-streamdown="code-block-body"] {
+  /* The white inner card from the example */
+  background: var(--background) !important;
+  border: 1px solid var(--border) !important;
+  border-radius: 0.375rem !important;
+  padding: 0.75rem !important; /* Reduced for tighter code bounding */
+  margin-top: 0 !important;
+}
+
+/* Neutralize aggressive prose styles that break Streamdown's inner <pre> */
+.prose pre {
+  background: transparent !important;
+  color: inherit !important;
+  margin: 0 !important;
+  padding: 0 !important;
+  border-radius: 0 !important;
+  border: none !important;
+}
+
+.prose code {
+  background: transparent !important;
+  font-weight: 400 !important;
+}
+
+.prose code::before,
+.prose code::after {
+  content: none !important;
+}
+
+/* Pass the shiki tokens down properly */
+[data-streamdown="code-block-body"] span[style*="--shiki-light"] {
+  color: var(--shiki-light) !important;
+}
+
+.dark [data-streamdown="code-block-body"] span[style*="--shiki-dark"] {
+  color: var(--shiki-dark) !important;
+}
